added a contributor stats view for mobilizations
