+++ public/js/contribute/mobilization/load.js
@@ -117,6 +117,14 @@ function DOMLoad() {
     .on('change.toggle', function () {
       togglePadLimit(this);
     });
+
+  // FOR STATS VIEW
+  d3.selectAll('tr.top-level')
+  .on('click', function () {
+    const { iso3 } = this.dataset
+    d3.select(this).toggleClass('open')
+    d3.selectAll(`tr.small.${iso3}`).toggleClass('hide')
+  })
 }
 
 if (document.readyState === 'loading') {
