+++ public/scss/_browse.lg.scss
@@ -1737,6 +1737,7 @@ div.browse {
                       small.total {
                         display: inline-block;
                         font-weight: normal;
+                        text-decoration: none !important;
                       }
                     }
                   }
