+++ public/scss/_browse.m.scss
@@ -1695,6 +1695,7 @@ div.browse {
                       small.total {
                         display: inline-block;
                         font-weight: normal;
+                        text-decoration: none !important;
                       }
                     }
                   }
