+++ public/scss/_browse.sm.scss
@@ -1680,6 +1680,7 @@ div.browse {
                         small.total {
                           display: inline-block;
                           font-weight: normal;
+                          text-decoration: none !important;
                         }
                       }
                     }
