+++ public/scss/_browse.xl.scss
@@ -1717,6 +1717,7 @@ div.browse {
                       small.total {
                         display: inline-block;
                         font-weight: normal;
+                        text-decoration: none !important;
                       }
                     }
                   }
