+++ public/scss/_browse.xs.scss
@@ -1876,6 +1876,7 @@ div.browse {
                         small.total {
                           display: inline-block;
                           font-weight: normal;
+                          text-decoration: none !important;
                         }
                       }
                     }
