+++ public/scss/_mobilize.lg.scss
@@ -890,3 +890,96 @@ main#mobilize-existing {
 // footer {
 
 // }
+
+table.stats {
+	width: 100%;
+	table-layout: fixed;
+	margin-bottom: 60px;
+	border-spacing: 0;
+
+	thead {
+    tr {
+      font-size: $t-mid;
+
+      th {
+        padding: 15px 30px;
+      }
+    }
+	}
+	tbody {
+		color: $c-mid-grey;
+
+		tr {
+      &.top-level {
+        td {
+          cursor: pointer;
+        }
+      }
+
+      &:hover {
+  			background-color: $c-light-3;
+  		}
+      &.open {
+        color: $c-dark-grey;
+
+        td {
+          border-bottom: 1px solid $c-light-grey;
+        }
+      }
+      &.small {
+        font-size: $t-mid-small;
+
+        & > td {
+          padding: 5px 30px;
+        }
+
+        &:nth-child(even) {
+          background: repeating-linear-gradient(
+            45deg,
+            transparent,
+            transparent 3px,
+            rgba(150, 150, 150, 0.1) 3px,
+            rgba(150, 150, 150, 0.1) 6px
+          );
+        }
+        &:nth-child(odd) {
+          background: repeating-linear-gradient(
+            -45deg,
+            transparent,
+            transparent 3px,
+            rgba(150, 150, 150, 0.1) 3px,
+            rgba(150, 150, 150, 0.1) 6px
+          );
+        }
+      }
+  		&.error {
+
+        &:nth-child(even) {
+    			background: repeating-linear-gradient(
+            45deg,
+            #fff,
+            #fff 3px,
+            rgba(165, 30, 65, 0.1) 3px,
+            rgba(165, 30, 65, 0.1) 6px
+          );
+        }
+        &:nth-child(odd) {
+          background: repeating-linear-gradient(
+            -45deg,
+            #fff,
+            #fff 3px,
+            rgba(165, 30, 65, 0.1) 3px,
+            rgba(165, 30, 65, 0.1) 6px
+          );
+        }
+  		}
+
+      td {
+        padding: 5px 15px;
+        &.center {
+          text-align: center;
+        }
+      }
+    }
+	}
+}
