+++ routes/contribute/mobilization/index.js
@@ -1,0 +1,3 @@
+exports.load = require('./load/')
+exports.render = require('./render.js')
+exports.authorization = require('./authorization.js')
