+++ routes/contribute/mobilization/load/index.js
@@ -1,0 +1,2 @@
+exports.data = require('./data.js')
+exports.stats = require('./stats.js')
