+++ routes/index.js
@@ -246,7 +246,7 @@ exports.dispatch.contribute = require('./contribute/')
 
 // exports.render.import = require('./import/').render
 
-exports.dispatch.mobilize = require('./mobilize/')
+// exports.dispatch.mobilize = require('./mobilize/')
 
 /* =============================================================== */
 /* =========================== BROWSE ============================ */
