+++ views/contribute/mobilization/stats.ejs
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<%
+	// CREATE ALIASES
+	const language = locals?.metadata?.page?.language
+%>
+<head>
+<%- include ('../../partials/head.ejs') %>
+
+<script type='module' src='/js/contribute/mobilization/load.js'></script>
+
+</head>
+
+<body>
+<%- include ('../../partials/navigate/menu-languages.ejs') %>
+
+<div class='main-content'>
+<header>
+	<div class='inner'>
+		<%- include ('../../partials/navigate/sidebar.ejs') %>
+	</div>
+</header>
+
+<div class="browse google-translate-attr exploration-info">
+	<% if (locals.metadata.page.referer?.includes('browse/mobilizations')) { %>
+		<a href='<%- locals.metadata.page.referer %>' class='return google-translate-attr' data-vocab='back to browse'></a>
+	<% } else { %>
+		<a href='<%- `/${language}/browse/mobilizations/ongoing` %>' class='return google-translate-attr' data-vocab='back to browse'></a>
+	<% } %>
+
+	<table class='stats'>
+		<thead>
+			<tr>
+				<!-- TO DO: TRANSLATE -->
+				<th data-vocab='country'></th>
+				<th># of participants</th>
+				<th># of contributions</th>
+			</tr>
+		</thead>
+		<tbody>
+		<% locals.participants?.forEach(d => { 
+			const submissions = contributions.find(c => c.iso3 === d.iso3)
+		%>
+			<tr class='top-level <%= d.iso3 %><% if ((submissions?.count ?? 0) === 0) { %> error<% } %>' data-iso3='<%= d.iso3 %>'>
+				<td><%= d.country || "Global" %></td>
+				<td class='center'><%= d.count %></td>
+				<td class='center'><%= submissions?.count ?? 0 %></td>
+			</tr>
+			<% d.values?.forEach(c => { 
+				const individualSubmissions = submissions?.values?.filter(b => b.owner === c.owner) %>
+			%>
+				<tr class='small <%= d.iso3 %> hide'>
+					<td><%= c.ownername %></td>
+					<td></td>
+					<td class='center'><%= individualSubmissions?.length ?? 0 %></td>
+				</tr>
+			<% }) %>
+		<% }) %>
+		</tbody>
+	</table>
+</div>
+
+<div class='sidebar lg'></div>
+</div>
+
+<footer>
+<%- include ('../../partials/footer.ejs') %>
+</footer>
+</main>
+
+</body>
+</html>
